+++ libraries/USB/examples/USBHostMouse/USBHostMouse.ino
@@ -43,8 +43,8 @@ void loop() {
   USBHost.task();
 
   if (USBHostMouse.mounted() && USBHostMouse.available()) {
-    int8_t dx = USBHostMouse.getX();
-    int8_t dy = USBHostMouse.getY();
+    int16_t dx = USBHostMouse.getX();
+    int16_t dy = USBHostMouse.getY();
     uint8_t btns = USBHostMouse.getButtons();
     int8_t wh = USBHostMouse.getWheel();
